Added `raw` subdirectory
