+++ src/lazy/binary/mod.rs
@@ -1,3 +1,8 @@
 mod encoded_value;
 pub mod immutable_buffer;
-pub mod lazy_raw_reader;
+pub mod lazy_reader;
+pub mod lazy_system_reader;
+pub mod raw;
+
+#[cfg(test)]
+mod test_utilities;
