+++ src/lazy/binary/raw/lazy_raw_value.rs
@@ -1,0 +1,396 @@
+use crate::binary::int::DecodedInt;
+use crate::binary::uint::DecodedUInt;
+use crate::lazy::binary::encoded_value::EncodedValue;
+use crate::lazy::binary::immutable_buffer::ImmutableBuffer;
+use crate::lazy::binary::raw::lazy_raw_sequence::LazyRawSequence;
+use crate::lazy::binary::raw::lazy_raw_struct::LazyRawStruct;
+use crate::lazy::binary::raw::raw_annotations_iterator::RawAnnotationsIterator;
+use crate::lazy::raw_value_ref::RawValueRef;
+use crate::result::{decoding_error, decoding_error_raw, incomplete_data_error};
+use crate::types::SymbolId;
+use crate::{Decimal, IonResult, IonType, RawSymbolTokenRef, Timestamp};
+use bytes::{BigEndian, ByteOrder};
+use std::fmt::{Debug, Formatter};
+use std::{fmt, mem};
+
+#[derive(Clone)]
+pub struct LazyRawValue<'data> {
+    pub(crate) encoded_value: EncodedValue,
+    pub(crate) input: ImmutableBuffer<'data>,
+}
+
+impl<'a> Debug for LazyRawValue<'a> {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        write!(
+            f,
+            "LazyRawValue {{\n  val={:?},\n  buf={:?}\n}}\n",
+            self.encoded_value, self.input
+        )
+    }
+}
+
+type ValueParseResult<'a> = IonResult<RawValueRef<'a>>;
+
+impl<'data> LazyRawValue<'data> {
+    pub fn ion_type(&self) -> IonType {
+        self.encoded_value.ion_type()
+    }
+
+    fn has_annotations(&self) -> bool {
+        self.encoded_value.has_annotations()
+    }
+
+    fn annotations_sequence(&self) -> ImmutableBuffer<'data> {
+        let offset_and_length = self
+            .encoded_value
+            .annotations_sequence_offset()
+            .map(|offset| {
+                (
+                    offset,
+                    self.encoded_value.annotations_sequence_length().unwrap(),
+                )
+            });
+        let (sequence_offset, sequence_length) = match offset_and_length {
+            None => return self.input.slice(0, 0),
+            Some(offset_and_length) => offset_and_length,
+        };
+        let local_sequence_offset = sequence_offset - self.input.offset();
+
+        self.input.slice(local_sequence_offset, sequence_length)
+    }
+
+    pub fn annotations(&self) -> RawAnnotationsIterator<'data> {
+        RawAnnotationsIterator::new(self.annotations_sequence())
+    }
+
+    // Return the part of the input buffer immediately following this encoded value
+    fn consume_bytes(&self) -> IonResult<ImmutableBuffer<'data>> {
+        let total_length = self.encoded_value.total_length();
+        if self.input.len() < total_length {
+            eprintln!("[consume_bytes] Incomplete {:?}", self);
+            return incomplete_data_error(
+                "only part of the current value is in the buffer",
+                self.input.offset(),
+            );
+        }
+        Ok(self.input.consume(total_length))
+    }
+
+    pub fn read(&self) -> ValueParseResult {
+        if self.encoded_value.header().is_null() {
+            let raw_value_ref = RawValueRef::Null(self.ion_type());
+            return Ok(raw_value_ref);
+        }
+
+        match self.ion_type() {
+            IonType::Null => unreachable!("all null types handled above"),
+            IonType::Bool => self.read_bool(),
+            IonType::Int => self.read_int(),
+            IonType::Float => self.read_float(),
+            IonType::Decimal => self.read_decimal(),
+            IonType::Timestamp => self.read_timestamp(),
+            IonType::Symbol => self.read_symbol(),
+            IonType::String => self.read_string(),
+            IonType::Clob => self.read_clob(),
+            IonType::Blob => self.read_blob(),
+            IonType::List => self.read_list(),
+            IonType::SExp => self.read_sexp(),
+            IonType::Struct => self.read_struct(),
+        }
+    }
+
+    // Can return Err if there aren't enough bytes available
+    fn value_body(&self) -> IonResult<&'data [u8]> {
+        let value_total_length = self.encoded_value.total_length();
+        if self.input.len() < value_total_length {
+            eprintln!("[value_body] Incomplete {:?}", self);
+            return incomplete_data_error(
+                "only part of the requested value is available in the buffer",
+                self.input.offset(),
+            );
+        }
+        let value_body_length = self.encoded_value.value_length();
+        let value_offset = value_total_length - value_body_length;
+        Ok(self.input.bytes_range(value_offset, value_body_length))
+    }
+
+    pub(crate) fn available_body(&self) -> ImmutableBuffer<'data> {
+        let value_total_length = self.encoded_value.total_length();
+        let value_body_length = self.encoded_value.value_length();
+        let value_offset = value_total_length - value_body_length;
+
+        let bytes_needed = std::cmp::min(self.input.len() - value_offset, value_body_length);
+        let buffer_slice = self.input.slice(value_offset, bytes_needed);
+        buffer_slice
+    }
+
+    pub(crate) fn field_name(&self) -> Option<SymbolId> {
+        self.encoded_value.field_id
+    }
+
+    fn read_bool(&self) -> ValueParseResult {
+        debug_assert!(self.encoded_value.ion_type() == IonType::Bool);
+        let representation = self.encoded_value.header().length_code;
+        let value = match representation {
+            0 => false,
+            1 => true,
+            invalid => {
+                return decoding_error(format!(
+                    "found a boolean value with an illegal representation (must be 0 or 1): {}",
+                    invalid
+                ))
+            }
+        };
+        Ok(RawValueRef::Bool(value))
+    }
+
+    fn read_int(&self) -> ValueParseResult {
+        debug_assert!(self.encoded_value.ion_type() == IonType::Int);
+        // `value_body()` returns a buffer starting at the body of the value.
+        // It also confirms that the entire value is in the buffer.
+        let uint_bytes = self.value_body()?;
+        let magnitude: Int = if uint_bytes.len() <= mem::size_of::<u64>() {
+            DecodedUInt::small_uint_from_slice(uint_bytes).into()
+        } else {
+            DecodedUInt::big_uint_from_slice(uint_bytes).into()
+        };
+
+        use crate::binary::type_code::IonTypeCode::*;
+        use crate::Int;
+        use num_traits::Zero;
+        let value = match (self.encoded_value.header.ion_type_code, magnitude) {
+            (PositiveInteger, integer) => integer,
+            (NegativeInteger, integer) if integer.is_zero() => {
+                return decoding_error("found a negative integer (typecode=3) with a value of 0");
+            }
+            (NegativeInteger, integer) => -integer,
+            _itc => return decoding_error("unexpected ion type code"),
+        };
+        Ok(RawValueRef::Int(value))
+    }
+
+    fn read_float(&self) -> ValueParseResult {
+        debug_assert!(self.encoded_value.ion_type() == IonType::Float);
+        let ieee_bytes = self.value_body()?;
+        let number_of_bytes = self.encoded_value.value_length();
+        let value = match number_of_bytes {
+            0 => 0f64,
+            4 => f64::from(BigEndian::read_f32(ieee_bytes)),
+            8 => BigEndian::read_f64(ieee_bytes),
+            _ => return decoding_error("encountered a float with an illegal length"),
+        };
+        Ok(RawValueRef::Float(value))
+    }
+
+    fn read_decimal(&self) -> ValueParseResult {
+        debug_assert!(self.encoded_value.ion_type() == IonType::Decimal);
+
+        if self.encoded_value.value_length() == 0 {
+            return Ok(RawValueRef::Decimal(Decimal::new(0i32, 0i64)));
+        }
+
+        let (exponent_var_int, remaining) = self.input.read_var_int()?;
+        let coefficient_size_in_bytes =
+            self.encoded_value.value_length() - exponent_var_int.size_in_bytes();
+
+        let exponent = exponent_var_int.value();
+        let (coefficient, _remaining) = remaining.read_int(coefficient_size_in_bytes)?;
+
+        if coefficient.is_negative_zero() {
+            return Ok(RawValueRef::Decimal(Decimal::negative_zero_with_exponent(
+                exponent,
+            )));
+        }
+
+        Ok(RawValueRef::Decimal(Decimal::new(coefficient, exponent)))
+    }
+
+    fn read_timestamp(&self) -> ValueParseResult {
+        debug_assert!(self.encoded_value.ion_type() == IonType::Timestamp);
+
+        let input = ImmutableBuffer::new(self.value_body()?);
+
+        let (offset, input) = input.read_var_int()?;
+        let is_known_offset = !offset.is_negative_zero();
+        let offset_minutes = offset.value() as i32;
+        let (year_var_uint, input) = input.read_var_uint()?;
+        let year = year_var_uint.value() as u32;
+
+        // Year precision
+
+        let builder = Timestamp::with_year(year);
+        if input.is_empty() {
+            let timestamp = builder.build()?;
+            return Ok(RawValueRef::Timestamp(timestamp));
+        }
+
+        // Month precision
+
+        let (month_var_uint, input) = input.read_var_uint()?;
+        let month = month_var_uint.value() as u32;
+        let builder = builder.with_month(month);
+        if input.is_empty() {
+            let timestamp = builder.build()?;
+            return Ok(RawValueRef::Timestamp(timestamp));
+        }
+
+        // Day precision
+
+        let (day_var_uint, input) = input.read_var_uint()?;
+        let day = day_var_uint.value() as u32;
+        let builder = builder.with_day(day);
+        if input.is_empty() {
+            let timestamp = builder.build()?;
+            return Ok(RawValueRef::Timestamp(timestamp));
+        }
+
+        // Hour-and-minute precision
+
+        let (hour_var_uint, input) = input.read_var_uint()?;
+        let hour = hour_var_uint.value() as u32;
+        if input.is_empty() {
+            return decoding_error("timestamps with an hour must also specify a minute");
+        }
+        let (minute_var_uint, input) = input.read_var_uint()?;
+        let minute = minute_var_uint.value() as u32;
+        let builder = builder.with_hour_and_minute(hour, minute);
+        if input.is_empty() {
+            let timestamp = if is_known_offset {
+                builder.build_utc_fields_at_offset(offset_minutes)
+            } else {
+                builder.build_at_unknown_offset()
+            }?;
+            return Ok(RawValueRef::Timestamp(timestamp));
+        }
+
+        // Second precision
+
+        let (second_var_uint, input) = input.read_var_uint()?;
+        let second = second_var_uint.value() as u32;
+        let builder = builder.with_second(second);
+        if input.is_empty() {
+            let timestamp = if is_known_offset {
+                builder.build_utc_fields_at_offset(offset_minutes)
+            } else {
+                builder.build_at_unknown_offset()
+            }?;
+            return Ok(RawValueRef::Timestamp(timestamp));
+        }
+
+        // Fractional second precision
+
+        let (subsecond_exponent_var_uint, input) = input.read_var_int()?;
+        let subsecond_exponent = subsecond_exponent_var_uint.value();
+        // The remaining bytes represent the coefficient.
+        let coefficient_size_in_bytes = self.encoded_value.value_length() - input.offset();
+        let (subsecond_coefficient, _input) = if coefficient_size_in_bytes == 0 {
+            (DecodedInt::zero(), input)
+        } else {
+            input.read_int(coefficient_size_in_bytes)?
+        };
+
+        let builder = builder
+            .with_fractional_seconds(Decimal::new(subsecond_coefficient, subsecond_exponent));
+        let timestamp = if is_known_offset {
+            builder.build_utc_fields_at_offset(offset_minutes)
+        } else {
+            builder.build_at_unknown_offset()
+        }?;
+
+        Ok(RawValueRef::Timestamp(timestamp))
+    }
+
+    /// If the reader is currently positioned on a symbol value, parses that value into a `SymbolId`.
+    pub fn read_symbol_id(&self) -> IonResult<SymbolId> {
+        debug_assert!(self.encoded_value.ion_type() == IonType::Symbol);
+        let uint_bytes = self.value_body()?;
+        if uint_bytes.len() > mem::size_of::<usize>() {
+            return decoding_error("found a symbol ID that was too large to fit in a usize");
+        }
+        let magnitude = DecodedUInt::small_uint_from_slice(uint_bytes);
+        // This cast is safe because we've confirmed the value was small enough to fit in a usize.
+        Ok(magnitude as usize)
+    }
+
+    pub fn read_symbol(&self) -> ValueParseResult {
+        debug_assert!(self.encoded_value.ion_type() == IonType::Symbol);
+        self.read_symbol_id()
+            .map(|sid| RawValueRef::Symbol(RawSymbolTokenRef::SymbolId(sid)))
+    }
+
+    /// If the reader is currently positioned on a string, returns a [&str] containing its text.
+    fn read_string(&self) -> ValueParseResult {
+        debug_assert!(self.encoded_value.ion_type() == IonType::String);
+        let raw_bytes = self.value_body()?;
+        let text = std::str::from_utf8(raw_bytes)
+            .map_err(|_| decoding_error_raw("found a string with invalid utf-8 data"))?;
+        Ok(RawValueRef::String(text))
+    }
+
+    fn read_blob(&self) -> ValueParseResult {
+        debug_assert!(self.encoded_value.ion_type() == IonType::Blob);
+        let bytes = self.value_body()?;
+        Ok(RawValueRef::Blob(bytes))
+    }
+
+    fn read_clob(&self) -> ValueParseResult {
+        debug_assert!(self.encoded_value.ion_type() == IonType::Clob);
+        let bytes = self.value_body()?;
+        Ok(RawValueRef::Clob(bytes))
+    }
+
+    fn read_sexp(&self) -> ValueParseResult {
+        debug_assert!(self.encoded_value.ion_type() == IonType::SExp);
+        let lazy_value = LazyRawValue {
+            encoded_value: self.encoded_value,
+            input: self.input,
+        };
+        let lazy_sequence = LazyRawSequence { value: lazy_value };
+        Ok(RawValueRef::SExp(lazy_sequence))
+    }
+
+    fn read_list(&self) -> ValueParseResult {
+        debug_assert!(self.encoded_value.ion_type() == IonType::List);
+        let lazy_value = LazyRawValue {
+            encoded_value: self.encoded_value,
+            input: self.input,
+        };
+        let lazy_sequence = LazyRawSequence { value: lazy_value };
+        Ok(RawValueRef::List(lazy_sequence))
+    }
+
+    fn read_struct(&self) -> ValueParseResult {
+        debug_assert!(self.encoded_value.ion_type() == IonType::Struct);
+        let lazy_value = LazyRawValue {
+            encoded_value: self.encoded_value,
+            input: self.input,
+        };
+        let lazy_struct = LazyRawStruct { value: lazy_value };
+        Ok(RawValueRef::Struct(lazy_struct))
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::lazy::binary::raw::lazy_raw_reader::LazyRawBinaryReader;
+    use crate::lazy::binary::test_utilities::to_binary_ion;
+    use crate::IonResult;
+
+    #[test]
+    fn annotations_sequence() -> IonResult<()> {
+        let data = &to_binary_ion(
+            r#"
+            $ion_symbol_table::{symbols: ["foo"]}
+            foo // binary writer will omit the symtab if we don't use a symbol 
+        "#,
+        )?;
+        let mut reader = LazyRawBinaryReader::new(data);
+        let _ivm = reader.next()?.expect_ivm()?;
+        let value = reader.next()?.expect_value()?;
+        let annotations_sequence = value.annotations_sequence();
+        assert_eq!(annotations_sequence.len(), 1);
+        assert_eq!(annotations_sequence.offset(), 6);
+        assert_eq!(annotations_sequence.bytes()[0], 0x83u8); // 0x83 == $3 == $ion_symbol_table
+        Ok(())
+    }
+}
