+++ src/lazy/mod.rs
@@ -1,6 +1,4 @@
 pub mod binary;
-pub mod lazy_reader;
-pub mod lazy_system_reader;
 pub mod raw_stream_item;
 pub mod raw_value_ref;
 pub mod system_stream_item;
