+++ tests/element_test_vectors.rs
@@ -5,7 +5,7 @@ use ion_rs::element::writer::{ElementWriter, Format, TextKind};
 use ion_rs::element::{Element, Sequence};
 use ion_rs::result::{decoding_error, IonError, IonResult};
 use ion_rs::IonData;
-use ion_rs::{BinaryWriterBuilder, IonWriter, Reader, TextWriterBuilder};
+use ion_rs::{BinaryWriterBuilder, IonWriter, TextWriterBuilder};
 
 use std::fs::read;
 use std::path::MAIN_SEPARATOR as PATH_SEPARATOR;
@@ -31,9 +31,6 @@ fn contains_path(paths: &[&str], file_name: &str) -> bool {
 // Statically defined arrays containing test file paths to skip in various contexts.
 type SkipList = &'static [&'static str];
 
-// TODO: This trait exists solely for the purposes of this integration test. It allows us to paper
-// over the differences in the ion-c and native Rust ElementWriter implementations.
-// Once all of the tests are passing, we should work to unify their APIs.
 trait RoundTrip {
     /// Encodes `elements` to a buffer in the specified Ion `format` and then reads them back into
     /// a `Vec<Element>` using the provided reader.
@@ -68,13 +65,15 @@ fn serialize(format: Format, elements: &[Element]) -> IonResult<Vec<u8>> {
 }
 
 trait ElementApi {
+    type ElementReader<'a>: ElementReader;
+
     fn global_skip_list() -> SkipList;
     fn read_one_equivs_skip_list() -> SkipList;
     fn round_trip_skip_list() -> SkipList;
     fn equivs_skip_list() -> SkipList;
     fn non_equivs_skip_list() -> SkipList;
 
-    fn make_reader(data: &[u8]) -> IonResult<Reader<'_>>;
+    fn make_reader(data: &[u8]) -> IonResult<Self::ElementReader<'_>>;
 
     /// Asserts the given elements can be round-tripped and equivalent, then returns the new elements.
     fn assert_round_trip(
@@ -427,6 +426,12 @@ mod native_element_tests {
     struct NativeElementApi;
 
     impl ElementApi for NativeElementApi {
+        type ElementReader<'a> = Reader<'a>;
+
+        fn make_reader(data: &[u8]) -> IonResult<Reader<'_>> {
+            ReaderBuilder::default().build(data)
+        }
+
         fn global_skip_list() -> SkipList {
             &[
                 // The binary reader does not check whether nested values are longer than their
@@ -492,10 +497,6 @@ mod native_element_tests {
 
         fn non_equivs_skip_list() -> SkipList {
             &[]
-        }
-
-        fn make_reader(data: &[u8]) -> IonResult<Reader<'_>> {
-            ReaderBuilder::default().build(data)
         }
     }
 
@@ -545,6 +546,24 @@ mod non_blocking_native_element_tests {
     struct NonBlockingNativeElementApi;
 
     impl ElementApi for NonBlockingNativeElementApi {
+        type ElementReader<'a> = Reader<'a>;
+
+        fn make_reader(data: &[u8]) -> IonResult<Self::ElementReader<'_>> {
+            // These imports are visible as a temporary workaround.
+            // See: https://github.com/amazon-ion/ion-rust/issues/484
+            use ion_rs::binary::constants::v1_0::IVM;
+            use ion_rs::reader::integration_testing::new_reader;
+            // If the data is binary, create a non-blocking binary reader.
+            if data.starts_with(&IVM) {
+                let raw_reader = RawBinaryReader::new(data);
+                Ok(new_reader(raw_reader))
+            } else {
+                // Otherwise, create a non-blocking text reader
+                let raw_reader = RawTextReader::new(data);
+                Ok(new_reader(raw_reader))
+            }
+        }
+
         fn global_skip_list() -> SkipList {
             &[
                 // The binary reader does not check whether nested values are longer than their
@@ -610,22 +629,6 @@ mod non_blocking_native_element_tests {
 
         fn non_equivs_skip_list() -> SkipList {
             &[]
-        }
-
-        fn make_reader(data: &[u8]) -> IonResult<Reader<'_>> {
-            // These imports are visible as a temporary workaround.
-            // See: https://github.com/amazon-ion/ion-rust/issues/484
-            use ion_rs::binary::constants::v1_0::IVM;
-            use ion_rs::reader::integration_testing::new_reader;
-            // If the data is binary, create a non-blocking binary reader.
-            if data.starts_with(&IVM) {
-                let raw_reader = RawBinaryReader::new(data);
-                Ok(new_reader(raw_reader))
-            } else {
-                // Otherwise, create a non-blocking text reader
-                let raw_reader = RawTextReader::new(data);
-                Ok(new_reader(raw_reader))
-            }
         }
     }
